move code to src/runtime/file_utils.h
